Update to v5.0.3, and match work done in https://github.com/Asana/codez/pull/233007.
This updates to v5.0.3, which is the most recent NPM version of fibers. Notably, the most recent version of fibers at https://github.com/laverdet/node-fibers appears to be 5.0.2, which I don't understand.

The main change to v5.0.3 is to remove certain logging. The main change in https://github.com/Asana/codez/pull/233007 was to add linux builds against the Node 18, which are included at s3://asana-oss-cache/node-fibers/fibers-5.0.3.tar.gz, but not in this repository.