Remove setupAsyncHacks
These hacks were added to address https://github.com/laverdet/node-fibers/issues/357, which was an issue in Node 8.

The hack works by clearing the async context stack when a fiber runs, and restoring it when the fiber yields. Unfortunately, the restoration isn't complete. At a minimum, it doesn't correctly restore the stack of async resources.

But also, the hack depends on async_wrap, which is now deprecated, and when the hack is removed, the original issues no longer reproduces, and unit tests pass in Asana. So as far as I can tell, the hack is no longer needed in node 18.

Test Plan:
Run the example code from https://github.com/laverdet/node-fibers/issues/357 to try to reproduce the issue in node 18. See that the issue doesn't reproduce.
(Note that this was in Asana's codebase, rather than running pure node)